FIXED: Alt-tab broke interaction with board
The only purpose for tracking the current key was to check which
individual modifier key was pressed. Alt-tabbing out made it so
that no key release event happened after the key press event,
which meant the tracking still showed an active keyboard modifier.
Instead, just fetch the currently tracked state of the keyboard
modifiders from the application.